test_data_generator/requirements.txt
google-adk
faker
exrex
python-dotenv
